+++ packages/orq-rc/src/lib/config.ts
@@ -37,6 +37,10 @@ export type SDKOptions = {
    */
   serverURL?: string | undefined;
   /**
+   * Allows overriding the default user agent used by the SDK
+   */
+  userAgent?: string | undefined;
+  /**
    * Allows overriding the default retry config used by the SDK
    */
   retryConfig?: RetryConfig;
@@ -64,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.7.0-rc.1",
-  genVersion: "2.610.0",
-  userAgent: "speakeasy-sdk/typescript 3.7.0-rc.1 2.610.0 2.0 @orq-ai/node",
+  sdkVersion: "3.7.0-rc.2",
+  genVersion: "2.616.1",
+  userAgent: "speakeasy-sdk/typescript 3.7.0-rc.2 2.616.1 2.0 @orq-ai/node",
 } as const;