+++ packages/orq-rc/RELEASES.md
@@ -1869,3 +1869,13 @@ Based on:
 - [typescript v3.7.0-rc.1] packages/orq-rc
 ### Releases
 - [NPM v3.7.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/3.7.0-rc.1 - packages/orq-rc
+
+## 2025-05-30 11:02:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.554.1 (2.616.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.7.0-rc.2] packages/orq-rc
+### Releases
+- [NPM v3.7.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/3.7.0-rc.2 - packages/orq-rc